+++ packages/google/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ai-sdk/google
 
+## 0.0.14
+
+### Patch Changes
+
+- df47a34: feat (provider/google): add models/gemini-1.5-flash-latest to list of supported models
+
+## 0.0.13
+
+### Patch Changes
+
+- 9a7ec8b: feat (provider/google): add token usage information
+
 ## 0.0.12
 
 ### Patch Changes